fix fifth pass of clippy lints
